+++ stac_api/api/extensions/__init__.py
@@ -1,9 +1,17 @@
 """stac_api.api.extensions"""
-from .context import ContextExtension  # noqa
+from .context import ContextExtension
+from .fields import FieldsExtension
+from .query import QueryExtension
+from .sort import SortExtension
+from .tiles import TilesExtension
+from .transaction import BulkTransactionExtension, TransactionExtension
 
-# from .extension import ApiExtension  # noqa
-from .fields import FieldsExtension  # noqa
-from .query import QueryExtension  # noqa
-from .sort import SortExtension  # noqa
-from .tiles import TilesExtension  # noqa
-from .transaction import TransactionExtension  # noqa
+__all__ = (
+    "ContextExtension",
+    "FieldsExtension",
+    "QueryExtension",
+    "SortExtension",
+    "TilesExtension",
+    "TransactionExtension",
+    "BulkTransactionExtension",
+)