+++ stac_api/api/extensions/fields.py
@@ -1,29 +1,31 @@
 """fields extension"""
-from dataclasses import dataclass, field
 from typing import Set
 
+import attr
 from fastapi import FastAPI
 
 from stac_api.api.extensions.extension import ApiExtension
 
 
-@dataclass
+@attr.s
 class FieldsExtension(ApiExtension):
     """
     stac-api fields extension
     (https://github.com/radiantearth/stac-api-spec/tree/master/extensions/fields)
     """
 
-    default_includes: Set[str] = field(
-        default_factory=lambda: {
-            "id",
-            "type",
-            "geometry",
-            "bbox",
-            "links",
-            "assets",
-            "properties.datetime",
-        }
+    default_includes: Set[str] = attr.ib(
+        default=attr.Factory(
+            lambda: {
+                "id",
+                "type",
+                "geometry",
+                "bbox",
+                "links",
+                "assets",
+                "properties.datetime",
+            }
+        )
     )
 
     def register(self, app: FastAPI) -> None: