+++ stac_api/clients/postgres/transactions.py
@@ -1,112 +1,141 @@
 """transactions extension client"""
 
+import json
 import logging
-from dataclasses import dataclass
-from typing import Type, Union
+from typing import Dict, Optional, Type
 
-from stac_api import errors
-from stac_api.clients.base import BaseTransactionsClient
-from stac_api.clients.postgres.base import PostgresClient
+import attr
+
+from stac_api.clients.base import BaseBulkTransactionsClient, BaseTransactionsClient
+from stac_api.clients.postgres.session import Session
+from stac_api.errors import NotFoundError
 from stac_api.models import database, schemas
 
 logger = logging.getLogger(__name__)
 
 
-@dataclass
-class TransactionsClient(PostgresClient, BaseTransactionsClient):
+@attr.s
+class TransactionsClient(BaseTransactionsClient):
     """Transactions extension specific CRUD operations"""
 
-    table: Type[database.Collection] = database.Collection
-    item_table: Type[database.Item] = database.Item
-
-    @property
-    def collection_table(self):
-        """alias for `self.table` # TODO: Figure out a better way to do this"""
-        return self.table
-
-    def _create(
-        self,
-        model: Union[schemas.Collection, schemas.Item],
-        table: Union[Type[database.Collection], Type[database.Item]],
-    ) -> Union[database.Collection, database.Item]:
-        """Create a single record"""
-        try:
-            self.lookup_id(model.id, table=table)
-            error_message = f"Row {model.id} already exists"
-            logger.error(error_message, exc_info=True)
-            raise errors.ConflictError(error_message)
-        except errors.NotFoundError:
-            row_data = table.from_schema(model)
-            self.writer_session.add(row_data)
-            self.commit()
-            return row_data
-
-    def _update(
-        self,
-        model: Union[schemas.Collection, schemas.Item],
-        table: Union[Type[database.Collection], Type[database.Item]],
-    ) -> Union[database.Collection, database.Item]:
-        """Create a single record if it does not exist or update an existing record"""
-        try:
-            query = self.lookup_id(model.id, table=table)
-            update_data = table.get_database_model(model)
-            # SQLAlchemy orm updates don't seem to like geoalchemy types
-            update_data.pop("geometry", None)
-            query.update(update_data)
-            self.commit()
-            return table.from_schema(model)
-        except errors.NotFoundError:
-            row_data = table.from_schema(model)
-            self.writer_session.add(row_data)
-            self.commit()
-            return row_data
-
-    def _delete(
-        self, item_id: str, table: Union[Type[database.Collection], Type[database.Item]]
-    ) -> Union[database.Collection, database.Item]:
-        """Delete a single record"""
-        query = self.lookup_id(item_id, table=table)
-        row_data = query.first()
-        query.delete()
-        self.commit()
-        return row_data
+    session: Session = attr.ib(default=attr.Factory(Session.create_from_env))
+    collection_table: Type[database.Collection] = attr.ib(default=database.Collection)
+    item_table: Type[database.Item] = attr.ib(default=database.Item)
 
     def create_item(self, model: schemas.Item, **kwargs) -> schemas.Item:
-        """Create an item"""
-        obj = self._create(model, table=self.item_table)
-        obj.base_url = str(kwargs["request"].base_url)
-        return schemas.Item.from_orm(obj)
-
-    def update_item(self, model: schemas.Item, **kwargs) -> schemas.Item:
-        """Update an item"""
-        obj = self._update(model, table=self.item_table)
-        obj.base_url = str(kwargs["request"].base_url)
-        return schemas.Item.from_orm(obj)
-
-    def delete_item(self, id: str, **kwargs) -> schemas.Item:
-        """Delete an item"""
-        obj = self._delete(id, table=self.item_table)
-        obj.base_url = str(kwargs["request"].base_url)
-        return schemas.Item.from_orm(obj)
+        """create item"""
+        data = self.item_table.from_schema(model)
+        with self.session.writer.context_session() as session:
+            session.add(data)
+            data.base_url = str(kwargs["request"].base_url)
+            return schemas.Item.from_orm(data)
 
     def create_collection(
         self, model: schemas.Collection, **kwargs
     ) -> schemas.Collection:
-        """Create a collection"""
-        obj = self._create(model, table=self.collection_table)
-        obj.base_url = str(kwargs["request"].base_url)
-        return schemas.Collection.from_orm(obj)
+        """create collection"""
+        data = self.collection_table.from_schema(model)
+        with self.session.writer.context_session() as session:
+            session.add(data)
+            data.base_url = str(kwargs["request"].base_url)
+            return schemas.Collection.from_orm(data)
+
+    def update_item(self, model: schemas.Item, **kwargs) -> schemas.Item:
+        """update item"""
+        with self.session.reader.context_session() as session:
+            query = session.query(self.item_table).filter(
+                self.item_table.id == model.id
+            )
+            if not query.scalar():
+                raise NotFoundError(f"Item {model.id} not found")
+            # SQLAlchemy orm updates don't seem to like geoalchemy types
+            data = self.item_table.get_database_model(model)
+            data.pop("geometry", None)
+            query.update(data)
+
+            response = self.item_table.from_schema(model)
+            response.base_url = str(kwargs["request"].base_url)
+            return schemas.Item.from_orm(response)
+        return model
 
     def update_collection(
         self, model: schemas.Collection, **kwargs
     ) -> schemas.Collection:
-        """Update a collection"""
-        obj = self._update(model, table=self.collection_table)
-        obj.base_url = str(kwargs["request"].base_url)
-        return schemas.Collection.from_orm(obj)
+        """update collection"""
+        with self.session.reader.context_session() as session:
+            query = session.query(self.collection_table).filter(
+                self.collection_table.id == model.id
+            )
+            if not query.scalar():
+                raise NotFoundError(f"Item {model.id} not found")
+            # SQLAlchemy orm updates don't seem to like geoalchemy types
+            data = self.collection_table.get_database_model(model)
+            data.pop("geometry", None)
+            query.update(data)
+        return model
+
+    def delete_item(self, id: str, **kwargs) -> schemas.Item:
+        """delete item"""
+        with self.session.writer.context_session() as session:
+            query = session.query(self.item_table).filter(self.item_table.id == id)
+            data = query.first()
+            if not data:
+                raise NotFoundError(f"Item {id} not found")
+            query.delete()
+            data.base_url = str(kwargs["request"].base_url)
+            return schemas.Item.from_orm(data)
 
     def delete_collection(self, id: str, **kwargs) -> schemas.Collection:
-        """Delete a collection"""
-        obj = self._delete(id, table=self.collection_table)
-        obj.base_url = str(kwargs["request"].base_url)
-        return schemas.Collection.from_orm(obj)
+        """delete collection"""
+        with self.session.writer.context_session() as session:
+            query = session.query(self.collection_table).filter(
+                self.collection_table.id == id
+            )
+            data = query.first()
+            if not data:
+                raise NotFoundError(f"Collection {id} not found")
+            query.delete()
+            data.base_url = str(kwargs["request"].base_url)
+            return schemas.Collection.from_orm(data)
+
+
+@attr.s
+class BulkTransactionsClient(BaseBulkTransactionsClient):
+    """postgres bulk transactions"""
+
+    session: Session = attr.ib(default=attr.Factory(Session.create_from_env))
+    debug: bool = attr.ib(default=False)
+
+    def __attrs_post_init__(self):
+        """create sqlalchemy engine"""
+        self.engine = self.session.writer.cached_engine
+
+    @staticmethod
+    def _preprocess_item(item: schemas.Item) -> Dict:
+        """
+        preprocess items to match data model
+        # TODO: dedup with GetterDict logic (ref #58)
+        """
+        item = item.dict(exclude_none=True)
+        item["geometry"] = json.dumps(item["geometry"])
+        item["collection_id"] = item.pop("collection")
+        item["datetime"] = item["properties"].pop("datetime")
+        return item
+
+    def bulk_item_insert(
+        self, items: schemas.Items, chunk_size: Optional[int] = None, **kwargs
+    ) -> str:
+        """
+        bulk item insertion using sqlalchemy core
+        https://docs.sqlalchemy.org/en/13/faq/performance.html#i-m-inserting-400-000-rows-with-the-orm-and-it-s-really-slow
+        """
+        # Use items.items because schemas.Items is a model with an items key
+        processed_items = [self._preprocess_item(item) for item in items.items]
+        return_msg = f"Successfully added {len(processed_items)} items."
+        if chunk_size:
+            for chunk in self._chunks(processed_items, chunk_size):
+                self.engine.execute(database.Item.__table__.insert(), chunk)
+            return return_msg
+
+        self.engine.execute(database.Item.__table__.insert(), processed_items)
+        return return_msg