+++ stac_api/models/schemas.py
@@ -8,14 +8,11 @@ from types import DynamicClassAttribute
 from typing import Any, Callable, Dict, List, Optional, Set, Union
 
 import sqlalchemy as sa
+from geojson_pydantic.geometries import Polygon
+from pydantic import BaseModel, Field, ValidationError, root_validator
+from pydantic.error_wrappers import ErrorWrapper
 from shapely.geometry import Polygon as ShapelyPolygon
 from shapely.geometry import shape
-
-from geojson_pydantic.geometries import Polygon
-from pydantic import Field, ValidationError, root_validator
-from pydantic.error_wrappers import ErrorWrapper
-from stac_api import config
-from stac_api.models.decompose import CollectionGetter, ItemGetter
 from stac_pydantic import Collection as CollectionBase
 from stac_pydantic import Item as ItemBase
 from stac_pydantic.api import Search
@@ -23,6 +20,9 @@ from stac_pydantic.api.extensions.fields import FieldsExtension as FieldsBase
 from stac_pydantic.api.search import DATETIME_RFC339
 from stac_pydantic.shared import Link
 from stac_pydantic.utils import AutoValueEnum
+
+from stac_api import config
+from stac_api.models.decompose import CollectionGetter, ItemGetter
 
 # Be careful: https://github.com/samuelcolvin/pydantic/issues/1423#issuecomment-642797287
 NumType = Union[float, int]
@@ -69,6 +69,7 @@ class Queryables(str, AutoValueEnum):
     minzoom = "cog:minzoom"
     maxzoom = "cog:maxzoom"
     dtype = "cog:dtype"
+    foo = "foo"
 
 
 @dataclass
@@ -165,6 +166,12 @@ class Item(ItemBase):
         use_enum_values = True
         orm_mode = True
         getter_dict = ItemGetter
+
+
+class Items(BaseModel):
+    """Items model"""
+
+    items: List[Item]
 
 
 class STACSearch(Search):